Problemas/Modulo #6/requirements.txt
Flask==2.3.3
redis==4.6.0
python-dotenv==1.0.0